[Sbuild::Build] invoke debsign post-build
Passing -k to dpkg-buildpackage involves the chroot having access to key
material.  Invoking debsign post-build outside of the chroot seems to make much
more sense.  It might be interactive at this point but at least for the
buildd.d.o case that's not going to happen, so that's ignored in this design.