Decouple chroot from host in terms of filesystem access
This is useful to allow sbuild backends where the chroot filesystem
cannot be directly accessed by the host, either because the host has no
permission to access the chroot or because the chroot is inside a
virtual machine, a container or on a remote system. This in turn allows
backends like adt-virt-qemu, adt-virt-ssh, adt-virt-lxc or backends
based on user namespaces to be added.

In course of the decoupling all pieces of code where:

 - the test operators (-d, -f, -s) were used
 - perl modules like File::Temp or File::Copy were used
 - perl functions like mkdir, open, close, opendir, readdir were used
 - external commands pointing to paths inside the chroot were used like
   apt-ftparchive

had to be replaced by invocations of a program that knows how to execute
processes inside the chroot.

To make this transition easier, many new convenience functions have been
added to lib/Sbuild/Chroot.pm like:

 - mktemp
 - copy_to_chroot
 - copy_from_chroot
 - get_read_file_handle
 - read_file
 - get_write_file_handle
 - read_command
 - write_file
 - write_command
 - rename
 - mkdir
 - test_directory
 - test_regular_file
 - test_regular_file_readable
 - test_symlink
 - unlink
 - chmod
 - chown

Since this also means that nothing in Build.pm or ResolverBase.pm needs
to know about the chroot path (because all modifications of the chroot
content have to go through Chroot.pm) the function strip_chroot_path is
not necessary anymore and with it the 'Location' and 'Chroot Dir'
configuration variables are only very rarely used.

FIXME: the %SBUILD_CHROOT_DIR variable for external command hooks is now
       not always useful anymore. Maybe it should be removed?

FIXME: a new variable for external command hooks should contain the
       command necessary to execute processes inside the chroot.

Also, with this commit, we create the Packages, Sources and Release file
inside the chroot.  We cannot use apt-ftparchive as this is not
available inside the chroot.  Apt-ftparchive outside the chroot might
not have access to the files inside the chroot for reasons explained
above. The only alternative would've been to set up the archive outside
the chroot using apt-ftparchive and to then copy Packages, Sources and
Release into the chroot.  We do not do this to avoid copying files from
and to the chroot.  At the same time doing it like this has the
advantage to have less dependencies of sbuild itself (no apt-ftparchive
needed).  The disadvantage of doing it this way is that we now have to
maintain our own code creating the Release file which might break in the
future.